Fix typings.
